Switch to f64

+++ src/alloy/alloy_types/bismuth_bronze.rs
@@ -44,8 +44,8 @@ impl private::AlloyType for BismuthBronze {
     }
 
     fn check_own_ranges_contains(
-        percentages: &[BaseMetal<f32>],
-    ) -> Result<Box<[BaseMetal<f32>]>, AlloyError> {
+        percentages: &[BaseMetal<f64>],
+    ) -> Result<Box<[BaseMetal<f64>]>, AlloyError> {
         let (mut seen_copper, mut seen_zinc, mut seen_bismuth) = (false, false, false);
         let mut reorder = [Copper(0.0), Copper(0.0), Copper(0.0)];
         for p in percentages {
