+++ src/alloy/alloy_types/cupronickel.rs
@@ -41,8 +41,8 @@ impl private::AlloyType for Cupronickel {
     }
 
     fn check_own_ranges_contains(
-        percentages: &[BaseMetal<f32>],
-    ) -> Result<Box<[BaseMetal<f32>]>, AlloyError> {
+        percentages: &[BaseMetal<f64>],
+    ) -> Result<Box<[BaseMetal<f64>]>, AlloyError> {
         let (mut seen_copper, mut seen_nickel) = (false, false);
         let mut reorder = [Copper(0.0), Copper(0.0)];
         for p in percentages {
