+++ src/alloy/alloy_types/lead_solder.rs
@@ -40,8 +40,8 @@ impl private::AlloyType for LeadSolder {
     }
 
     fn check_own_ranges_contains(
-        percentages: &[BaseMetal<f32>],
-    ) -> Result<Box<[BaseMetal<f32>]>, AlloyError> {
+        percentages: &[BaseMetal<f64>],
+    ) -> Result<Box<[BaseMetal<f64>]>, AlloyError> {
         let (mut seen_tin, mut seen_lead) = (false, false);
         let mut reorder = [Copper(0.0), Copper(0.0)];
         for p in percentages {
